backend/requirements.txt
fastapi
uvicorn
python-dotenv

# LLM
openai
langchain

# Resume parsing
pypdf
python-docx

# NLP
spacy
scikit-learn
numpy

# Utilities
tqdm
